refactor: 상품 상세페이지에 hydration boundary 적용 (#323)
