+++ apps/web/app/(header)/auction/[shortId]/page.tsx
@@ -1,8 +1,9 @@
 import AuctionDetailPageContent from '@/features/auction/detail/ui/AuctionDetailPageContent';
 import React from 'react';
 
-const AuctionDetailPage = ({ params }: { params: Promise<{ shortId: string }> }) => {
-  return <AuctionDetailPageContent params={params} />;
+const AuctionDetailPage = async ({ params }: { params: Promise<{ shortId: string }> }) => {
+  const { shortId } = await params;
+  return <AuctionDetailPageContent shortId={shortId} />;
 };
 
 export default AuctionDetailPage;
